+++ py/desispec/qa/qa_exposure.py
@@ -46,6 +46,7 @@ class QA_Exposure(object):
         self.night = night
         self.specprod_dir = specprod_dir
         self.flavor = flavor
+        self.meta = {}
 
         if in_data is None:
             self.data = dict(flavor=self.flavor, expid=self.expid,
@@ -84,6 +85,17 @@ class QA_Exposure(object):
 
         # Figure
         qa_plots.exposure_fluxcalib(outfil, self.data)
+
+    def load_meta(self, frame_meta):
+        """ Load meta info from input Frame meta
+        Args:
+            frame_meta:
+        """
+        desi_params = read_params()
+        for key in desi_params['frame_meta']:
+            if key in ['CAMERA']:  # Frame specific
+                continue
+            self.meta[key] = frame_meta[key]
 
     def load_qa_data(self, remove=False):
         """ Load the QA data files for a given exposure (currently yaml)